ffmpeg/qsv: move VP9-8 Encoder Test classes to lib
