+++ test/ffmpeg-qsv/encode/vp9.py
@@ -6,34 +6,11 @@
 
 from ....lib import *
 from ....lib.ffmpeg.qsv.util import *
-from ....lib.ffmpeg.qsv.encoder import EncoderTest
+from ....lib.ffmpeg.qsv.encoder import VP9_8EncoderLPTest
 
 spec = load_test_spec("vp9", "encode", "8bit")
 
-@slash.requires(*have_ffmpeg_encoder("vp9_qsv"))
-@slash.requires(*have_ffmpeg_decoder("vp9_qsv"))
-class VP9EncoderBaseTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      codec     = "vp9",
-      ffencoder = "vp9_qsv",
-      ffdecoder = "vp9_qsv",
-    )
-
-  def get_file_ext(self):
-    return "ivf"
-
-@slash.requires(*platform.have_caps("vdenc", "vp9_8"))
-class VP9EncoderLPTest(VP9EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps      = platform.get_caps("vdenc", "vp9_8"),
-      lowpower  = 1,
-    )
-
-class cqp_lp(VP9EncoderLPTest):
+class cqp_lp(VP9_8EncoderLPTest):
   def init(self, tspec, case, ipmode, qp, quality, slices):
     vars(self).update(tspec[case].copy())
     vars(self).update(
@@ -50,7 +27,7 @@ class cqp_lp(VP9EncoderLPTest):
     self.init(spec, case, ipmode, qp, quality, slices)
     self.encode()
 
-class cbr_lp(VP9EncoderLPTest):
+class cbr_lp(VP9_8EncoderLPTest):
   def init(self, tspec, case, gop, bitrate, fps, slices):
     vars(self).update(tspec[case].copy())
     vars(self).update(
@@ -70,7 +47,7 @@ class cbr_lp(VP9EncoderLPTest):
     self.init(spec, case, gop, bitrate, fps, slices)
     self.encode()
 
-class vbr_lp(VP9EncoderLPTest):
+class vbr_lp(VP9_8EncoderLPTest):
   def init(self, tspec, case, gop, bitrate, fps, slices, quality):
     vars(self).update(tspec[case].copy())
     vars(self).update(
